refactor(workshop): move reconciliation out of read endpoints
